feat!: handle different versions of the same resource
This repository has been using a very old version of aip-go, which still
contains the `aipregistry`. This functionality was removed because it
did not properly handle different versions of the same resource.

This commit updates the dependency on aip-go (and others) and rewrites
the plugin to handle "external" parent resources similar to other
plugins.

Previously a parent resources was searched for in *all* files in the
generate request, and if it encountered duplicates (i.e. different
versions) it would fail.

Now parent resources are only searched for in files in the same package.
To have the plugin generate resource names for resources outside the
current package, use "forward declarations" by adding a file level
resource definition (see the added file
`example/proto/einride/example/todo/v1/resources.proto`). This behaviour
matches other resource name generator plugins.